neuro/requirements.txt
# Зависимости нейросетевого модуля матчинга товаров
tensorflow>=2.15.0
numpy>=1.24.0
faiss-cpu>=1.7.4
tqdm>=4.65.0
